moved position

+++ workers/atlasbrowser.py
@@ -62,15 +62,16 @@ def generate_spatial(self, qcparams, **kwargs):
 
     #remove all files from the temp folder. To allevaite bugs being caused by figure folder being generated using old images.
     temp_path = Path(temp_dir).joinpath(root_dir, run_id, 'images')
-    for filename in os.listdir(temp_path):
-        file_path = os.path.join(temp_path, filename)
-        try:
-            if os.path.isfile(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print('Failed to delete ' + file_path + " due to " + e)
+    if os.path.exists(temp_path):
+        for filename in os.listdir(temp_path):
+            file_path = os.path.join(temp_path, filename)
+            try:
+                if os.path.isfile(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    shutil.rmtree(file_path)
+            except Exception as e:
+                print('Failed to delete ' + file_path + " due to " + e)
 
     ### source image path
     allFiles = [i for i in oldFiles if '.json' not in i and 'spatial' not in i]
